Experimental Changes for Grinder

+++ spec/generate_ec2_ssh_keys.rb
@@ -18,48 +18,55 @@ keys_file = File.join("..", "config", "cloud_variables", "ec2_keys.json")
 ssh_dir = File.join(File.expand_path("~"), ".ssh")
 rest_yaml = File.join(File.expand_path("~"), ".rest_connection", "rest_api_config.yaml")
 rest_settings = YAML::load(IO.read(rest_yaml))
-rest_settings[:ssh_keys] = []
-specific_key_file = File.join(ssh_dir, "specific_keys")
-specific_key_names = YAML::load(IO.read(specific_key_file)) if File.exists?(specific_key_file)
+rest_settings[:ssh_keys] = [] unless rest_settings[:ssh_keys]
+multicloud_key_file = File.join(ssh_dir, "api_user_key")
+multicloud_key_data = IO.read(multicloud_key_file) if File.exists?(multicloud_key_file)
 if File.exists?(keys_file)
   keys = JSON::parse(IO.read(keys_file))
 else
   keys = {}
 end
 
+api0_1 = false
+begin
+  api0_1 = Ec2SshKeyInternal.find_all
+rescue
+end
+
 cloud_ids.each { |cloud|
-  next if cloud == 0
-  if File.exists?(specific_key_file)
-    key_name = specific_key_names[:names][cloud - 1]
+  next if cloud == 0 # Not a valid cloud ID
+  next if keys["#{cloud}"] # We already have data for this cloud, skip
+  if File.exists?(multicloud_key_file)
+    key_name = "api_user_key"
   else
     if cloud <= 10
       key_name = "monkey-#{cloud}-#{ENV['RS_API_URL'].split("/").last}"
-    elsif cloud == 850
-      key_name = "publish-test"
     else
       key_name = "monkey-1-#{ENV['RS_API_URL'].split("/").last}"
     end
   end
   if cloud <= 10
-    found = Ec2SshKeyInternal.find_by_cloud_id("#{cloud}").select { |obj| obj.aws_key_name =~ /#{key_name}/ }.first
+    found = nil
+    if api0_1
+      found = Ec2SshKeyInternal.find_by_cloud_id("#{cloud}").select { |o| o.aws_key_name =~ /#{key_name}/ }.first
+    end
     k = (found ? found : Ec2SshKey.create('aws_key_name' => key_name, 'cloud_id' => "#{cloud}"))
     keys["#{cloud}"] = {"ec2_ssh_key_href" => k.href,
                         "parameters" =>
                           {"PRIVATE_SSH_KEY" => "key:#{key_name}:#{cloud}"}
                         }
+    # Generate Private Key Files
+    priv_key_file = File.join(ssh_dir, "monkey-cloud-#{cloud}")
+    File.open(priv_key_file, "w") { |f| f.write(k.aws_material) } unless File.exists?(priv_key_file)
   else
-    found = Ec2SshKeyInternal.find_by_cloud_id("1").select { |obj| obj.aws_key_name =~ /#{key_name}/ }.first
-    k = (found ? found : Ec2SshKey.create('aws_key_name' => key_name, 'cloud_id' => "1"))
-    keys["#{cloud}"] = {"parameters" =>
-                          {"PRIVATE_SSH_KEY" => "key:#{key_name}:1"}
-                        }
+    # Use API user's managed ssh key
+    puts "Using API user's managed ssh key"
+    priv_key_file = multicloud_key_file
   end
-  # Generate Private Key Files
-  priv_key_file = File.join(ssh_dir, "monkey-cloud-#{cloud}")
-  File.open(priv_key_file, "w") { |f| f.write(k.aws_material) }
+
   File.chmod(0700, priv_key_file)
   # Configure rest_connection config
-  rest_settings[:ssh_keys] << priv_key_file
+  rest_settings[:ssh_keys] << priv_key_file unless rest_settings[:ssh_keys].contains?(priv_key_file)
 }
 
 keys_out = keys.to_json(:indent => "  ",
